+++ packages/web3-core/src/web3_config.ts
@@ -25,7 +25,7 @@ import {
 } from 'web3-types';
 import { ConfigHardforkMismatchError, ConfigChainMismatchError } from 'web3-errors';
 import { isNullish, toHex } from 'web3-utils';
-import { TransactionTypeParser } from './types.js';
+import { CustomTransactionSchema, TransactionTypeParser } from './types.js';
 // eslint-disable-next-line import/no-cycle
 import { TransactionBuilder } from './web3_context.js';
 import { Web3EventEmitter } from './web3_event_emitter.js';
@@ -59,6 +59,7 @@ export interface Web3ConfigOptions {
 	};
 	transactionBuilder?: TransactionBuilder;
 	transactionTypeParser?: TransactionTypeParser;
+	customTransactionSchema?: CustomTransactionSchema;
 	defaultReturnFormat: DataFormat;
 }
 
@@ -101,6 +102,7 @@ export abstract class Web3Config
 		},
 		transactionBuilder: undefined,
 		transactionTypeParser: undefined,
+		customTransactionSchema: undefined,
 		defaultReturnFormat: DEFAULT_RETURN_FORMAT,
 	};
 
@@ -518,6 +520,15 @@ export abstract class Web3Config
 	public set transactionTypeParser(val) {
 		this._triggerConfigChange('transactionTypeParser', val);
 		this.config.transactionTypeParser = val;
+	}
+
+	public get customTransactionSchema(): CustomTransactionSchema | undefined {
+		return this.config.customTransactionSchema;
+	}
+
+	public set customTransactionSchema(schema: CustomTransactionSchema | undefined) {
+		this._triggerConfigChange('customTransactionSchema', schema);
+		this.config.customTransactionSchema = schema;
 	}
 
 	private _triggerConfigChange<K extends keyof Web3ConfigOptions>(