+++ packages/web3-core/test/unit/web3_config.test.ts
@@ -49,6 +49,7 @@ const defaultConfig = {
 	defaultReturnFormat: DEFAULT_RETURN_FORMAT,
 	transactionBuilder: undefined,
 	transactionTypeParser: undefined,
+	customTransactionSchema: undefined,
 };
 const setValue = {
 	string: 'newValue',