+++ packages/web3-eth/src/types.ts
@@ -37,6 +37,7 @@ import {
 	TransactionWithFromLocalWalletIndex,
 	TransactionWithToLocalWalletIndex,
 } from 'web3-types';
+import { Schema } from 'web3-validator';
 
 export type InternalTransaction = FormatType<Transaction, typeof ETH_DATA_FORMAT>;
 
@@ -105,3 +106,8 @@ export interface TransactionMiddleware {
 		options?: { [key: string]: unknown },
 	): Promise<TransactionMiddlewareData>;
 }
+
+export type CustomTransactionSchema = {
+	type: string;
+	properties: Record<string, Schema>;
+};