+++ packages/web3-eth-personal/CHANGELOG.md
@@ -147,4 +147,10 @@ Documentation:
 
 -   Dependencies updated
 
+## [4.1.0]
+
+### Changed
+
+-   Forwards the new `web3Context.config.customTransactionSchema` to `formatTransaction`
+
 ## [Unreleased]