+++ packages/web3-eth/CHANGELOG.md
@@ -269,4 +269,14 @@ Documentation:
 -   Adds transaction property to be an empty list rather than undefined when no transactions are included in the block (#7151)
 -   Change method `getTransactionReceipt` to not be casted as `TransactionReceipt` to give proper return type (#7159)
 
+## [4.9.0]
+
+### Changed
+
+-   Forwards the new `web3Context.config.customTransactionSchema` to `formatTransaction`
+
+### Added
+
+-   Adds the same `{transactionSchema?: ValidationSchemaInput}` that exists in `formatTransaction` to `validateTransactionForSigning`
+
 ## [Unreleased]